move inkscape to the publishing task
Prior to this commit, inkscape was part of the Apps task and eigher
installed via dnf or flatpak, depending on the flatpak parameter.
However, inkscape is a dependency for texlive-xput (part of
texlive-collection-pictures) and thus was installed via dnf regardless
of the flatpak parameter. By moving inkscape to the publishing task,
having two installations of it is avoided.